Don't read past the Name buffer if the symbol name exceeds the max length.
SymFromAddrW sets NameLen to the real length of the symbol, which can be
greater than MaxNameLen. Before this change, such a symbol would cause
the code to read past the end of the Name buffer into garbage. With this
change, the code reads only till the min of NameLen and (MaxNameLen - 1).

Fixes #59