Add a feature to symbolicate with `gimli-rs/addr2line`
This commit adds a new feature, "gimli", which enables symbolicating addresses
into file/line/function information using the `gimli-rs/addr2line` crate.

This is currently Linux only. We could enable this for macOS as well if we had a
way to find the dSYM bundle for a given shared library or executable file. That
involves looking for the LC_UUID load command in the mach-o file, and then using
spotlight APIs to find the corresponding bundle containing the DWARF debug
infos.

At least on Linux, this is one less C dependency. RIIR!!