Remove synchronization from dbghelp init
This will rely on external synchronization but we'll still keep an
internal counter to make sure we don't initialize more than once.
Performance optimizations for `Backtrace` will come in a later commit.

Additionally don't succeed initialization if `SymInitializeW` failed for
now, but we can catch it later if necessary in the future.