Load dbghelp.dll dynamically on Windows
This commit updates the Windows bindings for this library to load
dbghelp.dll dynamically instead of linking to it statically at compile
time. This matches what libstd does and in theory reduces the
dependencies of a binary application on Windows ensuring that
information related to backtraces is deferred until it's actually
needed.