Remove dependency on `winapi`
This is the next step towards moving the `backtrace` crate into the
standard library, removing the dependency on the `winapi` crate. The
standard library doesn't have access to `winapi`, so neither will the
`backtrace` crate!

Definitions are duplicated into this crate, but all verified against
`winapi` on CI to ensure they're correct.