Add support for libbacktrace on MinGW
This commit adds support to use libbacktrace for symbolication on MinGW
and switches to it by default since dbghelp doesn't really work on MinGW
anyway.

This largely involvse juggling a few dependencies around but
additionally adding support to the `libbacktrace.rs` module for loading
the filename of the current image on Windows.