Favor debuginfo function names instead of symbol table
This commit fixes an accidental regression with #185 where libbacktrace
was updated to match the standard library by always returning the symbol
table symbol. Turns out though this library was one-upping the standard
library by consulting debuginfo (what it was previously doing) because
debuginfo can have more accurate symbol information for inline
functions, for example.

The fix here is to always prefer the debuginfo function name, if
present, and otherwise fall back to the symbol table symbol if
necessary.

Closes #186