Don't adjust the IP with libunwind
This was first introduced with the very first implementation of
symbolication way back when in
rust-lang/rust@bdd31b38aaf320acc9dad5fa012ba2834348c1f8, but nowadays
it's not entirely clear that this is still necessary. In any case it's
best to have the adjustment done in symbolication code rather than the
return value of `Frame::ip`, so switch to returning an unadjusted
instruction address.

If this causes issues we can always have some offsets later on in the
return values!