gimli: Only iterate shared libraries once
This commit updates the gimli implementation to iterate over all shared
libraries only once, caching information about them to get processed
later. This information is usually static for the lifetime of the
program, and this improves the performance of resolving a backtrace
significantly in some situations (by up to 30%).

The organization of the gimli backend has been updated to reflect this
cache where `Cache` is now a dedicated type with methods hanging off it.
Some helper code had to be duplicated from the `findshlibs` crate to do
this entirely.

cc #232