gimli: Sort symbols on Linux
Instead of relying on lazy iteration, instead iterate over all symbols
up front and then sort them by address. We already do this on Windows
and it looks like symbol iteration is somewhat slow in benchmarks for
goblin. By sorting we should also be able to run through these lists
much more quickly!

Overall on top of #235 this yields another 30% speedup or so for
symbolication.

cc #232