gimli: Look at object dynamic symbol tables
This commit updates the gimli parsing code to consult the dynamic symbol
table for each object in addition to the normal symbol table. This also
starts filtering symbols in the same way libbacktrace does to ensure
symbolication doesn't go down the rabbit hole with non-function
symbols.

This overall removes the need for gimli to fall back to dladdr
functionally for now. In testing this speeds up benchmarks by 15% by
avoiding calling `dladdr` in libc.