Refactor Fuchsia printing support to all platforms
This commit refactors the support needed to print Fuchsia's backtraces
to support all platforms as well. This ideally moves out the need to
have different code for printing a backtrace in libstd and this crate,
and now both can share the same implementation!

The implementation should produce the same output on fuchsia, but for
other platforms it should match the current output of the standard
library, including the full/short configuration options.