Remove coresymbolication implementation (#316)
* Remove coresymbolication implementation

This isn't used anywhere by default any more and I don't think there's
any reason to enable it. This older implementation uses private APIs
which aren't supposed to be used. Additionally libbacktrace nowadays
should be workable enough and with gimli soon on the horizon that should
be a suitable default for macOS and iOS.

* Disable android emulator in CI