Handle compressed debug sections in ELF files (#344)
ELF files allow debug info sections to be compressed. The libbacktrace
backed supported these compressed sections, but the Gimli backend did
not. This commit adds that support to the Gimli backend.

In my tests (with the BFD linker, lld, and gold) these debug info
sections do not obey the alignment requirements that the object crate
expects for the gABI compression header, so this commit additionally
enables the "unaligned" feature in the upcoming version of the object
crate.

There is a bit of unsafe to ensure the lifetime of the decompressed
sections matches the lifetime of the mmap'd file. I don't think there is
a way around this unsafe code, unless we are willing to ditch Gimli's
EndianSlice for an (apparently slower) EndianReader backed by a
Cow<[u8]>.

Fix #342.