implement From<Frame> for BacktraceFrame (#420)
* implement From<Frame> for BacktraceFrame

There are situations where we can only capture raw `Frame` (example:
capturing from a signal handler), but it could still be that we can
process them later and are not fully nostd in the environment.

With a conversion from Frame to BacktraceFrame, this is trivial. But
without it, we can't really use the pretty printers as they are reliant
on BacktraceFrame, not Frame.

Fixes: #419

* add test

* fixes for miri

Every backtrace that miri generates - conversion or no conversion, has
totally different addresses. They all resolve to the same thing, though,
so test with that.

* fix another broken test

`cargo test --no-default-features --features std` fails because
frames are not resolved to symbols in that case. We'll just remove the
assert, rather than disabling the whole test on that case. At least the
code paths are stressed.