Avoid transmutes with `Context` (#429)
This commit re-consolidates transmutes around `Context` to just one
location to ensure that this is in as few places as possible. This
reorganizes the location of the `map_sup` variable and places it inside
of `Stash` instead which is a sort of auxiliary bag for extra
information that can be referenced anyway. Afterwards with some
refactoring an a new `Mapping::mk` function it's back to one transmute,
yay!