TEMPORARY: revert to bors@rust-lang.org for the merge commit email
We want Ferrocene's bors instance as the author email for merge commits,
but in the squashed repository there is no commit with that email yet,
causing CI to fail. This should be reverted as soon as a PR is merged
into the squashed repository.

This also changes the fetch depth to ensure we pick up a
bors@rust-lang.org commit.