allow releases of commits not present in this repository
Right now, the source code of Ferrocene 23.06.0 still lives in a private
repository, while the release process runs in ferrocene/ferrocene. Until
we open source 23.06.1, we need a way to publish releases for 23.06.0.

The release tooling is able to publish releases for 23.06.0 from the
public repository (if the full commit hash is provided), but the
resolution of the git ref into a commit hash would fail.

This adds a new option to skip resolving the git ref when starting a
manual release.