fix Cargo.lock autofix spuriously not triggering
Before this commit, the Cargo.lock automatic fix in pull-upstream would
only trigger occasionally in a nondeterministic manner.

That was caused by a weird interaction of `grep --quiet` and `set -o
pipefail`. When invoked with the `--quiet` flag, grep exits as soon as
the first match is found, closing the pipe. The command feeding into
grep though didn't handle well the pipe being closed, and exited with an
error whenever grep closed the pipe before it finished writing. Thanks
to `set -o pipefail`, this would result in the whole pipeline returning
1 even though grep returned 0.

This commit fixes the problem by always expecting the `grep` output, and
piping it into /dev/null not to show it to the user.