Switch the aarch64-unknown-none target to use rust-lld.
Using the bundled rust-lld as the linker (via it's ld.lld wrapper) means all users always use the same linker. This helps with the qualification efforts, rather than users using random versions of `ld`.

We don't go via the system C compiler on aarch64-unknown-none because we do not need system C libraries. We do go via the system C compiler on aarch64-unknown-ferrocenecoretest because that needs libstd to build the test runner.