Fix panic in backtrace symbolication on win7
Since https://github.com/rust-lang/backtrace-rs/pull/569 was merged,
symbolication of backtraces would panic on Windows 7, due to using
symbols that do not exist in the version of dbghelp.dll that ships there
(Windows 7 ships with dbghelp.dll version 6.1, but the symbols were only
added in version 6.2).

This commit checks for the presence of the newer symbols, and if they
are not found, falls back to the old resolution method.