Ferrocene Self Test: Execute sample programs
If you are compiling a sample binary, and you are running on the self-test host, then execute the binary and checkout the standard output is as expected.

This verifies that the compiler produces executable binaries, and that the linker is receiving the object files passed to the chosen C compiler (acting as the linker driver).