Add Functionality to ignore line in cov-report
Adding "COVR_EXCL_LINE" comment to the source code will ignore that line being
accounted for the code coverage report.

For example, in this code snippet the line is never reached,
So we might not be able to test this line of code.
In this case the keyword can be used to ignore it from the coverage report.

```rust
None => unsafe { hint::unreachable_unchecked() }, // COVR_EXCL_LINE

```

Co-authored-by: Atri Sarkar <atri.sarkar@thoughtworks.com>