store objects.inv into a separate directory
Before this commit, both the objects.inv produced during gathering and
the resulting HTML build were stored on the same directory. This worked
fine until now, but with the addition of --fresh it started to misbehave
regarding breadcrumbs.

The build process works like this:

* Start the build
* Copy breadcrumbs files
* Build intersphinx objects.inv, for each document:
  * Start the build
  * Invoke Sphinx
* Invoke Sphinx

The problem is, during "start the build" the directory was deleted, so
the intersphinx "start the build" would delete the files copied by "copy
breadcrumbs files". This commit switches the intersphinx objects.inv
building to a separate directory, so that the two don't interfere with
each other.