do not check status code of git grep
`git grep` (like any other grep) returns a non-zero exit code when no
file were found, and with check=True that would result in a Python
exception. This is not the behavior we want though, as not finding any
conflict marker is not an error.

This didn't show up before filtering directories was added because some
conflict markers were always found, they were just ignored by the rest
of the script and not displayed.