Fix deny(unused) of an unused import with SGX + Miri (rust-lang/backtrace-rs#581)
The `deny(unused)` use that used to be here is incompatible with building
this crate with Miri. The import in question is only used in a module
that is only declared when we are not `cfg(miri)`.

I tried to fix this by grouping items into modules so that I could apply
`cfg(not(miri))` to all the SGX code. One could also make the build work
by deleting the `#[deny(unused)]`, but that felt hacky.