remove channel sphinx substitution
Turns out we can't inject the channel name in our documentation! Our
document signature scheme depends on the content of the signature not to
change between the beta and stable branch (as we need to keep the same
document IDs and signatures between the two), while injecting the
channel implies the signatures would be invalidated.

Ferrocene-backport-of: #404