Mark `panic` as certified.
This commit is pretty invasive as it basically overrides `panic_fmt` so
it just calls `core::intrinsics::abort` unconditionally and ignores any
expression passed to it.

It also means that we can only certify `core` if `panic_immediate_abort`
is enabled, so this commit also adds it as a default feature.