s/Ferrocene Language Specification/FLS/ and update README
This works to replace "Ferrocene Language Specification" with "FLS"
across the document until we come up with our preferred name to
replace the FLS.

This includes removing the Ferrocene logo and replacing it with the
Rust Project logo in the sidebar.

This also updates the README to provide appropriate credit to Ferrous
Systems and provide the history of the FLS when it was the Ferrocene
Language Specification.