Move external shared Ferrocence extensions locally to FLS Project
Ferrocene has a shared resource package for their documentation and other sites that was used for the FLS: https://pypi.org/project/ferrocene-sphinx-shared-resources/. In order for us to make changes to the theme, logo, etc., I moved some of these extensions locally to the FLS Project since we wouldn't be able to make any changes to these things in that external package since it would affect other Ferrocene things too.

Now there are no dependencies on the external Ferrocene packages and the site builds as normal with them locally.