tooling: Replace `git branch -D && git checkout -b` with `checkout -B`
This gives it transactional semantics and makes error messages more
clear.

From the git man page for `git checkout`:

---

checkout:

If -B is given, <new-branch> is created if it doesn’t exist; otherwise, it is reset. This is the transactional equivalent of

    $ git branch -f <branch> [<start-point>]
    $ git checkout <branch>

that is to say, the branch is not reset/created unless "git checkout" is successful.

branch:

-f, --force
    Reset <branchname> to <startpoint>, even if <branchname> exists already. Without -f, git branch refuses to change an existing branch. In
    combination with -d (or --delete), allow deleting the branch irrespective of its merged status, or whether it even points to a valid
    commit. In combination with -m (or --move), allow renaming the branch even if the new branch name already exists, the same applies for -c
    (or --copy).