Split tidy into a separate CI job
Move tidy from the main "checks" job into a new "tidy" job. This gets
faster feedback for formatting failures, and simultaneously does not
require re-pushing a new commit just to run x fmt.

To make iteration times as fast as possible, this only clones the subset
of submodules required by tidy to run. It also contains a small fix to
`ferrocene/checkout-submodules.sh` to make it possible to run the script
locally; see the diff for details.