bootstrap: `PathSet::check` only considers `starts_with` for `--skip` flag
The original motivation for adding this `path.starts_with` check was to
allow things like `--skip=src/etc`. But it affects quite a lot more
things than just `--skip`; bootstrap is really not designed for the case
when multiple Steps match the same filter. For example, `x test src`
does ... something! Not sure what, but something!

Change `starts_with` to only affect `--skip`, not anything else.

The original motivation for this was to make it so that `x doc src/doc
--open` doesn't open a dozen different books, but I expect it to fix
various other steps in bootstrap as well.