bootstrap: flip-link is a target tool, not a bootstrap tool
In particular, we distribute it, which means it should be possible to
compile with stage2. This also allows us to cross-compile a flip-link
from e.g. a Darwin host to a Linux target; that's totally fine.

This commit isn't strictly necessary to run flip-link tests, it just
stood out to me while working on them.