Add myself and Ana as codeowners for bootstrap tests
The main motivation here is that bootstrap tests tend to hit a lot of
merge conflicts, and it's very easy to just `--bless` them away. I would
like to see how the conflicts are resolved and make sure that we aren't
testing the wrong behavior.

Ana is added as a backup in case I'm on vacation.