glossary: Unicode: refer to the Consortium's one, not ISO/IEC's
The Unicode Standard and ISO/IEC 10646 are different standards [1].

Since `fls_jpecw46eh061` refers to a Unicode Standard version, change
the glossary entry to point to the Consortium's standard instead.

This may also make it easier to refer to future versions since there may
be no published ISO/IEC standard for that version, at least for some time.

The description between "is" and "described" is taken from Chapter 1
("Introduction") of the Core Specification.

The version is not specified since it is already noted elsewhere (and, in
principle, it could be possible that different parts refer to different
versions -- if that is not desirable, it may make sense to move the
version here instead, like for the C standard one).

Fixes: https://github.com/rust-lang/fls/issues/155
Fixes: https://github.com/rust-lang/fls/issues/156
Link: https://www.unicode.org/faq/unicode_iso [1]
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>