qualification report: Infer host from test outcomes
Previously, the code hard-coded the host as x86_64-unknown-linux-gnu,
and it was simply impossible to generate `local` outcomes on any other
host.

This makes two changes:
- Test outcomes are only keyed by target, not by host. This will come in
  handy in the future when we plan to have aarch64 test runners.
- The host is infered from the tests run on that target, rather than
  being pre-determined beforehand. This fixes `local` outcomes, and
  should make it less work to have multiple hosts in the future.

Note that this still does not allow multiple hosts, which is enforced by
an assert in the `render-outcomes-template` directive.
It simply allows the (singular) host to be something other than x64 linux.