x doc: Unconditionally enable `--fresh`
Without this flag, bootstrap can sometimes serve stale and incorrect
artifacts for Sphinx documents. `--fresh` doesn't add that much compile time
and it causes a lot of confusion when it's missing, so just
unconditionally enable it.

If this really starts to be a problem in the future, we can readd it as
`--no-incremental` or something, but unsoundness should always be an
*opt-in*, not an opt-out.